First batch of code added
A lot of this code was developed outside of a repository.
I am moving it here now for the more focsed debugging and
refactoring.

The first files I am adding are for extracing a lung widnow CT image
from the DICOM files. Next up is the code for the matching algorithm.
We also need to refactor the saving code, the modeling code, and some
more stuff that I have overlooked.