Adding functionality to the model library
Added a function that will make the fully connected part of the model more programmatically.
We give it a list of the number of 'neurons' we want the model to have in each layer, and it uses the list to create teh layers all the way to the output layer.

Also added a function that abstracts away the dropout(activation(linear layer())) pattern that we were repeating all over the place.